+++ packages/gestalt/src/SelectList.js
@@ -5,13 +5,17 @@ import classnames from 'classnames';
 import Box from './Box.js';
 import formElement from './FormElement.css';
 import FormErrorMessage from './FormErrorMessage.js';
+import FormHelperText from './FormHelperText.js';
+import FormLabel from './FormLabel.js';
 import Icon from './Icon.js';
 import styles from './SelectList.css';
 
 type Props = {|
   errorMessage?: string,
   disabled?: boolean,
+  helperText?: string,
   id: string,
+  label?: string,
   name?: string,
   onChange: ({ event: SyntheticInputEvent<>, value: string }) => void,
   options: Array<{
@@ -32,7 +36,9 @@ export default class SelectList extends React.Component<Props, State> {
   static propTypes = {
     disabled: PropTypes.bool,
     errorMessage: PropTypes.string,
+    helperText: PropTypes.string,
     id: PropTypes.string.isRequired,
+    label: PropTypes.string,
     name: PropTypes.string,
     onChange: PropTypes.func.isRequired,
     options: PropTypes.arrayOf(
@@ -72,7 +78,9 @@ export default class SelectList extends React.Component<Props, State> {
     const {
       disabled,
       errorMessage,
+      helperText,
       id,
+      label,
       name,
       options,
       placeholder,
@@ -90,6 +98,7 @@ export default class SelectList extends React.Component<Props, State> {
 
     return (
       <Box>
+        {label && <FormLabel id={id} label={label} />}
         <Box
           color={disabled ? 'lightGray' : 'white'}
           dangerouslySetInlineStyle={{ __style: { borderRadius: 4 } }}
@@ -139,12 +148,10 @@ export default class SelectList extends React.Component<Props, State> {
             ))}
           </select>
         </Box>
-
-        {errorMessage && (
-          <Box marginTop={2}>
-            <FormErrorMessage id={id} text={errorMessage} />
-          </Box>
-        )}
+        {helperText && !errorMessage ? (
+          <FormHelperText text={helperText} />
+        ) : null}
+        {errorMessage && <FormErrorMessage id={id} text={errorMessage} />}
       </Box>
     );
   }